BugFix: If the user is not signed in redirect /browse to login page and vice-versa
