+++ netflix-gpt-swati/src/components/Body.js
@@ -1,35 +1,10 @@
-import React, { useEffect } from "react";
-import { addUser, removerUser } from "./utils/userSlice";
-
 import Browse from "./Browse";
 import Login from "./Login";
+import React from "react";
 import { RouterProvider } from "react-router-dom";
-import { auth } from "./utils/firebase";
 import { createBrowserRouter } from "react-router-dom";
-import { onAuthStateChanged } from "firebase/auth";
-import { useDispatch } from "react-redux";
 
 const Body = () => {
-  const dispatch = useDispatch();
-
-  useEffect(() => {
-    onAuthStateChanged(auth, (user) => {
-      if (user) {
-        const { uid, email, displayName, photoURL } = user;
-        dispatch(
-          addUser({
-            uid: uid,
-            email: email,
-            displayName: displayName,
-            photoURL: photoURL,
-          })
-        );
-      } else {
-        dispatch(removerUser());
-      }
-    });
-  }, []);
-
   const appRouter = createBrowserRouter([
     {
       path: "/",
@@ -40,6 +15,7 @@ const Body = () => {
       element: <Browse />,
     },
   ]);
+  
   return (
     <div>
       <RouterProvider router={appRouter} />
