+++ netflix-gpt-swati/src/components/Header.js
@@ -1,19 +1,44 @@
-import React from "react";
+import React, { useEffect } from "react";
+import { addUser, removerUser } from "./utils/userSlice";
+
 import { auth } from "./utils/firebase";
+import { onAuthStateChanged } from "firebase/auth";
 import { signOut } from "firebase/auth";
+import { useDispatch } from "react-redux";
 import { useNavigate } from 'react-router-dom';
 import { useSelector } from "react-redux";
 
 const Header = () => {
   const user = useSelector(store => store.user);
+
   const navigate = useNavigate();
+  const dispatch = useDispatch();
+
   const handleSignout = () => {
-    signOut(auth).then(() => {
-      navigate("/");
-    }).catch((error) => {
+    signOut(auth).then(() => {}).catch((error) => {
       navigate("/error");
     });
   }
+
+  useEffect(() => {
+    onAuthStateChanged(auth, (user) => {
+      if (user) {
+        const { uid, email, displayName, photoURL } = user;
+        dispatch(
+          addUser({
+            uid: uid,
+            email: email,
+            displayName: displayName,
+            photoURL: photoURL,
+          })
+        );
+        navigate("/browse");
+      } else {
+        dispatch(removerUser());
+        navigate("/")
+      }
+    });
+  }, []);
 
   return (
     <div className="absolute w-screen px-8 py-2 bg-gradient-to-b from-black z-10 flex justify-between">
